% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_functions.R
\name{build_imp_column_names}
\alias{build_imp_column_names}
\title{Takes SPSS Data that has a label format of "XXXX "Job Name" - "Something you don't want" and convert it the column name and strip everything except the "Job_Name"... It has to be in that format}
\usage{
build_imp_column_names(df, job_section)
}
\arguments{
\item{job_section}{name of job section}

\item{data}{this is the dataframe you want, needs LABELS from importing via Haven}
}
\value{
Dataframe with updated column names
}
\description{
Takes SPSS Data that has a label format of "XXXX "Job Name" - "Something you don't want" and convert it the column name and strip everything except the "Job_Name"... It has to be in that format
}
\examples{
build_imp_column_names(df,"job_step")
}
