% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_jtbd_cols.R
\name{build_sat_column_names}
\alias{build_sat_column_names}
\title{Rename satisfaction columns so they can be used in 'get_jtbd_scores'}
\usage{
build_sat_column_names(df, job_section)
}
\arguments{
\item{df}{a subsetted data frame that only includes the satisfaction columns from your dataset}

\item{job_section}{tA text string in parentheses; this is the grouping variable, most often the "job step"}
}
\value{
a data frame with duplicated columns with "sat__job_section." added to the front of the column name
}
\description{
Rename satisfaction columns so they can be used in 'get_jtbd_scores'
}
\examples{
#' build_sat_column_names(your_data_frame,"job_step_1")
build_sat_column_names(df.clean,"find_available_options")
}
