% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_imp_sat_data_cols.R
\name{convert_labels_to_row_names}
\alias{convert_labels_to_row_names}
\title{Takes SPSS Data that has a label format of "XXXX "Job Name" - "Something you don't want" and convert it the column name and strip everything except the "Job_Name"... It has to be in that format}
\usage{
convert_labels_to_row_names(df)
}
\arguments{
\item{data}{this is the dataframe you want, needs LABELS from importing via Haven}

\item{job_section}{name of job section}
}
\value{
Dataframe with updated column names
}
\description{
Takes SPSS Data that has a label format of "XXXX "Job Name" - "Something you don't want" and convert it the column name and strip everything except the "Job_Name"... It has to be in that format
}
\examples{
build_sat_column_names(data = "df_imp_cols", job_section = "Finding_A_Restaurant")
build_imp_column_names(data = "df_sat_cols", job_section = "Finding_A_Restaurant")
}
