% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_opp_scores.R
\name{get_imp_sat_opp_scores_total_population}
\alias{get_imp_sat_opp_scores_total_population}
\title{Computes the opportunity scores at the dataframe level}
\usage{
get_imp_sat_opp_scores_total_population(your_data_frame)
}
\arguments{
\item{your_data_frame}{input your cleaned dataframe that includes importance and satisfaction columns named in the following format:
imp__job_section.job_name
sat__job_section.job_name}
}
\value{
table of importance, satisfaction, and opportunity scores
}
\description{
Computes the opportunity scores at the dataframe level
}
\examples{
opp_scores <- get_imp_sat_opp_scores_total_population(your_data_frame)
}
