% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_jtbd_scores.R
\name{get_jtbd_scores}
\alias{get_jtbd_scores}
\title{Calculate the importance, satisfaction, and opportunity scores}
\usage{
get_jtbd_scores(your_data_frame, segment_var)
}
\arguments{
\item{your_data_frame}{your dataframe}

\item{segment_var}{this is the text string, within parentheses, that will be appended to each of the rows, we typically use "all"}
}
\value{
a data frame of scores
}
\description{
Calculate the importance, satisfaction, and opportunity scores
}
\examples{
#' get_jtbd_scores(df.clean,"all")
get_jtbd_scores(df.clean,"total_pop")
get_jtbd_scores(df,"young_adults")
}
