% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jtbd_scores.R
\name{jtbd_scores}
\alias{jtbd_scores}
\title{JTBD_SCORES}
\usage{
jtbd_scores(your_data_frame)
}
\arguments{
\item{your_data_frame}{}
}
\value{
a datframe of
}
\description{
JTBD_SCORES
}
\examples{
jtbd_scores("jtbd_importance_and_satisfaction_df")



}
