% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_spss_col_imp_sat.R
\name{rename_spss_col_imp_sat}
\alias{rename_spss_col_imp_sat}
\title{Title}
\usage{
rename_spss_col_imp_sat(
  df,
  unique_string_the_column_name_starts_with,
  job_section,
  imp_or_sat_string
)
}
\arguments{
\item{imp_or_sat_string}{}
}
\description{
Title
}
\examples{
df_sample <- rename_spss_col.imp_sat(df,"unique_column_prefix","job_section","imp_or_sat")


}
