% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_jtbd_cols.R
\name{build_imp_column_names}
\alias{build_imp_column_names}
\title{Rename importance columns so they can be used in 'get_jtbd_scores'}
\usage{
build_imp_column_names(df, job_section)
}
\arguments{
\item{df}{a subsetted data frame that only includes the importance columns from your dataset}

\item{job_section}{A text string in parentheses; this is the grouping variable, most often the "job step"}
}
\value{
a data frame with duplicated columns with "imp__job_section." added to the front of the column name
}
\description{
Rename importance columns so they can be used in 'get_jtbd_scores'
}
\examples{
build_imp_column_names(1,"find_available_options")
}
